+++ src/Compressor/Zip.php
@@ -9,6 +9,7 @@ class Zip extends AbstractCompressor
      */
     protected $config = [];
 
+
     /**
      * @param array $config
      *
@@ -18,6 +19,7 @@ class Zip extends AbstractCompressor
     {
         parent::__construct($config);
     }
+
 
     /**
      * @return void|bool
@@ -31,39 +33,77 @@ class Zip extends AbstractCompressor
         if (!preg_match('~\.zip$~ui', $dest)) {
             $dest .= '.zip';
         }
+        $dest_pathinfo = pathinfo($dest);
+        $dest_files = [];
 
         //
         $arguments = ['zip'];
         if (!empty($this->config['password'])) {
-            $arguments[] = '-P ' . $this->config['password'];
+            $arguments[] = '--password="' . $this->config['password'] . '"';
+        }
+        if (!empty($this->config['compress'])) {
+            $arguments[] = '-' . $this->config['compress'];
         }
         $arguments[] = '-r';
         $arguments[] = $dest;
         $arguments[] = $this->config['src'];
-        $arguments[] = $this->excludePrepare();
+        $arguments[] = $this->getExcludeArgs();
+        $arguments[] = '2>&1';
 
         //
         $command = join(' ', $arguments);
-        $result = shell_exec($command);
+        shell_exec($command);
+        unset($arguments, $command);
 
-        //
-        if (!file_exists($dest)) {
+        if (!empty($this->config['split'])) {
+            $arguments = [
+                'split',
+                '"' . $dest . '"',
+                '-b ' . $this->config['split'],
+                '-d',
+                '"' . $dest . '."',
+                // '2>&1',
+            ];
+            $command = join(' ', $arguments);
+            shell_exec($command);
+            unset($arguments, $command);
+
+            foreach (scandir($dest_pathinfo['dirname']) as $k) {
+                if (in_array($k, ['.', '..'], true)) {
+                    continue;
+                }
+                if (preg_match('~^' . preg_quote($dest_pathinfo['basename'], '~') . '\.[0-9]+~u', $k)) {
+                    $dest_files[] = $dest_pathinfo['dirname'] . '/' . $k;
+                }
+            }
+        }
+        if (empty($dest_files)) {
+            $dest_files[0] = $dest;
+        } elseif (count($dest_files) === 1 && md5_file($dest_files[0]) === md5_file($dest)) {
+            @unlink($dest_files[0]);
+            $dest_files[0] = $dest;
+        } else {
+            @unlink($dest);
+        }
+
+        if (!file_exists($dest_files[0])) {
             throw new \Exception("It was not possible to pack the source `{$this->config['src']}`.");
         }
-        $this->setFilePath($dest);
+        $this->setFiles($dest_files);
 
         return true;
     }
 
+
     /**
      * @return string
      */
-    protected function excludePrepare()
+    protected function getExcludeArgs()
     {
         $exclude = [];
         if (!empty($this->config['exclude'])) {
             foreach ($this->config['exclude'] as $v) {
-                if (substr($v, 0, 1) == '/') {
+                if (substr($v, 0, 1) === '/') {
                     $v = $this->config['src'] . $v;
                 }
                 $exclude[] = '-x ' . str_replace('*', '\*', $v);