+++ src/Storage/AbstractStorage.php
@@ -2,8 +2,14 @@
 
 namespace Gvozdb\Dumper\Storage;
 
+use Gvozdb\Dumper\Backup;
+
 abstract class AbstractStorage
 {
+    /**
+     * @var Backup $dumper
+     */
+    protected $dumper;
     /**
      * @var array $config
      */
@@ -13,19 +19,37 @@ abstract class AbstractStorage
      */
     public $message;
 
+
     /**
+     * @param Backup $dumper
      * @param array $config
      *
      * @throws \Exception
      */
-    public function __construct(array $config = [])
+    public function __construct(Backup $dumper, array $config = [])
     {
+        $this->dumper = $dumper;
+
         $this->config = array_merge([
             'path' => '',
         ], $config);
 
         $this->config['path'] = strftime(@$this->config['path'] ?: '');
+
+        $this->config['expires'] = is_array($this->config['expires'])
+            ? $this->config['expires']
+            : ['short_max_days' => $this->config['expires']];
+        $this->config['expires']['short_max_days'] = @$this->config['expires']['short_max_days'] ?: 1;
+        $this->config['expires'] = array_merge([
+            'short_step' => 1,
+            'long_max_days' => $this->config['expires']['short_max_days'],
+            'long_step' => 1,
+        ], $this->config['expires']);
+        $this->config['expires']['long_max_days'] = $this->config['expires']['long_max_days'] < $this->config['expires']['short_max_days']
+            ? $this->config['expires']['short_max_days']
+            : $this->config['expires']['long_max_days'];
     }
+
 
     /**
      * @return bool
@@ -41,6 +65,7 @@ abstract class AbstractStorage
         return true;
     }
 
+
     /**
      * @param $filepath
      *
@@ -48,10 +73,12 @@ abstract class AbstractStorage
      */
     abstract public function upload($filepath);
 
+
     /**
      * @return void|bool
      */
     abstract public function clean();
+
 
     /**
      * @return string