+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+
+## [0.3.0] - 2021-11-20
+
+* Added the ability to select the archiver (section `compressor` in the config)
+* Added the ability to specify short and long term storage of backups, as well as the step of obsolescence checking (section `expires` in config)
+* Added support for level of compression in `Compressor\Zip`
+* Added support for splitting the archive into files of a certain size in `Compressor\Zip`
+* Added progress bar when uploading files to `Storage\YandexDisk`
+* Added public method `Config\Load::setDefault`
+* Added public method `Backup::progressBar`
+* Fixed definition of the default prefix value for archive names
+* Moved to specify the archive password in `compressor.password` in the main config and in the user config
+* Properties `main.expires` and `storages.{Cloud}.expires` in the main config are declared deprecated
+* Property `main.archive_password` in the main config is declared deprecated
+* Property `archive_password` in the user config is declared deprecated
+
+
 ## [0.2.2] - 2021-05-05
 
 * Added storage support for single user
@@ -19,8 +36,8 @@
 * Added output of error messages
 * Added information messages output
 * Added recording of logs to file
-* Added sending of logs to mail
-* Added sending of logs to Telegram Chat
+* Added sending of logs to email
+* Added sending of logs to Telegram chat
 * Changed README.md